refactor: reuse getting dirname
