+++ src/fs/delete.js
@@ -1,10 +1,10 @@
 import { rm } from 'node:fs/promises';
 import path from 'node:path';
-import { getDir } from './utils.js';
+import { getDir } from '../utils.js';
 
 const remove = async () => {
   try {
-    await rm(path.join(getDir(), 'files','fileToRemove.txt'));
+    await rm(path.join(getDir(import.meta.url), 'files','fileToRemove.txt'));
   } catch {
     throw new Error('FS operation failed');
   }
