+++ src/fs/list.js
@@ -1,10 +1,10 @@
 import path from 'path';
-import { getDir } from './utils.js';
+import { getDir } from '../utils.js';
 import { readdir } from 'fs/promises';
 
 const list = async () => {
     try {
-		const res = await readdir(path.join(getDir(), 'files'),  { recursive: true});
+		const res = await readdir(path.join(getDir(import.meta.url), 'files'),  { recursive: true});
         console.log(res);
     } catch {
         throw new Error('FS operation failed');
