--- src/fs/utils.js
@@ -1,5 +1,0 @@
-import path from 'path';
-import { fileURLToPath } from 'node:url';
-
-
-export const getDir = () => path.dirname(fileURLToPath(import.meta.url));
