+++ src/streams/read.js
@@ -1,9 +1,9 @@
 import { createReadStream } from 'node:fs';
 import path from 'path';
-import { fileURLToPath } from 'node:url';
 import { stdout } from 'node:process';
+import { getDir } from '../utils';
 
-const dir = path.dirname(fileURLToPath(import.meta.url));
+const dir = getDir(import.meta.url);
 const pathToRead = path.join(dir, './files/fileToRead.txt' );
 const readStream = createReadStream(pathToRead, 'utf-8');
