+++ src/utils.js
@@ -1,0 +1,5 @@
+import { dirname } from 'node:path';
+import { fileURLToPath } from 'node:url';
+
+export const getDir= (url) => dirname(fileURLToPath(url));
+
