+++ src/zip/decompress.js
@@ -4,7 +4,7 @@ import { createUnzip } from 'node:zlib';
 import path from 'node:path';
 import { fileURLToPath } from 'node:url';
 
-const dir = path.dirname(fileURLToPath(import.meta.url));
+const dir = getDir(import.meta.url);
 
 const outputPath = path.join(dir, './files/fileToCompress.txt');
 const inputPath = path.join(dir, './files/archive.gz');
